Enemy Programming Begun
Building enemies is underway. The patterns are built, but there is an error that occurs when the program tries to run that Cargo.build did not detect. Currently unsure what is causing it, but just as the window opens, the program crashes; frustrating, but rust has proven to be nothing else.
Next step is to attempt to figure out what the error is. We will ask our teacher about it.