MAIN GAME FINISHED!
We now have the main game complete! Added sprites and AI movement to boot! Will need to double-check the requriements on the project, but the demo is ready for presentation. (the player can't even get out of the screen anymore!) after double-checking requirements, anything else is icing on the cake.